Finished
Added comments and changed the numbers from floats to integers

+++ hw4/part_2.cpp
@@ -4,11 +4,19 @@ using namespace std;
 
 int main()
 {
+    // First score
     float first_score;
+
+    // Second score
     float second_score;
+
+    // Third score
     float third_score;
+
+    // Fourth score
     float fourth_score;
 
+    // For loop over the range of 0 to 2
     for (int i = 0; i < 2; i++)
     {
         cout << "Type in the first score: ";
@@ -23,10 +31,12 @@ int main()
         cout << "Type in the fourth score: ";
         cin >> fourth_score;
 
-        float sum = first_score + second_score + third_score + fourth_score;
+        // Adds up all of the scores
+        int sum = first_score + second_score + third_score + fourth_score;
         cout << "Sum: " << sum << endl;
 
-        float average = sum / 4.0;
+        // Takes the average of the scores
+        int average = sum / 4.0;
         cout << "Average: " << average << endl;
     }
